+++ libft/geometry/quaternion/quat_helper.c
@@ -6,7 +6,7 @@
 /*   By: tischmid <tischmid@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/07 21:52:38 by tischmid          #+#    #+#             */
-/*   Updated: 2024/09/12 01:17:00 by tischmid         ###   ########.fr       */
+/*   Updated: 2024/09/19 08:29:09 by tischmid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ float	quat_len_squared(t_quat *quat)
 
 float	quat_len(t_quat *quat)
 {
-	return (sqrt(quat_len_squared(quat)));
+	return (sqrtf(quat_len_squared(quat)));
 }
 
 void	quat_print(t_quat *quat)
@@ -37,11 +37,11 @@ void	quat_print(t_quat *quat)
 }
 
 // potential speedup possible
-// 1 == sqrt(cos(degrees * PI / 360)**2 + length_squared_vec3(x, y, z));
+// 1 == sqrtf(cos(degrees * PI / 360)**2 + length_squared_vec3(x, y, z));
 // 1 == cos(degrees * PI / 360)**2 + length_squared_vec3(x, y, z);
 // 1 - cos(degrees * PI / 360)**2 == length_squared_vec3(x_new, y_new, z_new);
 // curr_length == length_squared_vec3(x_curr, y_curr, z_curr);
-// x_new / x_curr = sin(degrees * PI / 360) / sqrt(curr_length);
+// x_new / x_curr = sin(degrees * PI / 360) / sqrtf(curr_length);
 
 //todo later: maybe use cosf
 // expects a unit vector!!!
@@ -52,9 +52,9 @@ void	new_unit_quat(t_quat *quat, float degrees, t_vec3 *axis)
 	t_vec3	unit_axis;
 
 	copy_vec3(&unit_axis, axis);
-	sc_mult_vec3(&unit_axis, sin(degrees * PI / 360));
+	sc_mult_vec3(&unit_axis, sinf(degrees * PI / 360));
 	quat_copy(quat, &(t_quat){
-		.scalar = cos(degrees * PI / 360),
+		.scalar = cosf(degrees * PI / 360),
 		.x = unit_axis.x,
 		.y = unit_axis.y,
 		.z = unit_axis.z
